pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regvps-telegram-bot"
version = "0.1.0"
description = "Telegram-бот для start/stop/reboot VPS (Reg.ru CloudVPS / reglets API)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "python-telegram-bot>=21,<23",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "respx>=0.21",
  "ruff>=0.6",
]

[project.scripts]
vps-telegram = "vps_telegram_bot.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["vps_telegram_bot*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "RUF100"]
